chore: add preflight checks to pipelines (#520)
* chore: add image preflight checks to pipelines

Konflux does not have a way of introducing new tasks into our build
pipelines automatically. Instead, we need to do these things manually.
The image preflight checks are new as of a couple of months ago, and are
now being added to all new components as they are onboarded. This commit
brings the operator pipelines up to date with that addition.

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: actually reference the task bundle

Signed-off-by: Lance Ball <lball@redhat.com>

---------

Signed-off-by: Lance Ball <lball@redhat.com>